+++ src/cs3500/animator/model/SimpleRectangle.java
@@ -7,8 +7,8 @@ import java.awt.Color;
  */
 public class SimpleRectangle extends SimpleShape implements Shape  {
 
-  public SimpleRectangle(String name, double x, double y,
-                         Color color, double height, double width) {
+  public SimpleRectangle(String name, float x, float y,
+                         Color color, float height, float width) {
     super(name, x, y, color, height, width);
     this.type = ShapeType.RECTANGLE;
   }
